+++ features/step_definitions/general_steps/messages_steps.rb
@@ -1,3 +1,7 @@
+Given("I should see this error message {string} on this field {string}") do |expected_text, css|
+  validate_error_message(expected_text, css)
+end
+
 Given(/^I should see the error message "(.*?)"$/) do |expected_text|
   sleep 2
   p_tag_message.should eq(expected_text)