+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 All notable changes to this project will be documented in this file. This project does adhere to [Semantic Versioning](http://semver.org/)
 
 # 2.5.28 (9/25/18)
-- AWE-1860: Fixed some Wallet tests to work with the new nav bar
+- AWE-1860: Fix Wallet User Tests to work with the new Nav Bar
 
 # 2.5.27 (9/24/18)
 - AWE-1820: KBA: Edit Review Info